Revert "Typecheck"
