+++ .eslintrc.js
@@ -26,10 +26,7 @@ module.exports = {
     jest: true,
   },
   plugins: ['@typescript-eslint', 'react'],
-  ignorePatterns: [
-    '_this_is_virtual_fs_path_/$virtual/sync-requires.js',
-    'gatsby.js',
-  ],
+  ignorePatterns: ['_this_is_virtual_fs_path_/$virtual/sync-requires.js'],
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
