--- src/Hoc/hoc.types.tsx
@@ -1,6 +1,0 @@
-export interface WithLayoutProps {
-  // TODO: typecheck data
-  data?: any;
-  // data?: { [key: string]: string | { [key: string]: string } };
-  locale?: string;
-}
