+++ src/components/Pages/ProjectPageComponent.tsx
@@ -1,14 +1,11 @@
 // Libs
 import React from 'react';
 
-// Utils
-import { WithLayoutProps } from '../../Hoc/hoc.types';
-
 // Components
 import withLayout from '../../Hoc/PageWrapper/WithLayout';
 import SEO from '../seo';
 
-const ProjectPageComponent: React.FC<WithLayoutProps> = (props) => {
+const ProjectPageComponent: React.FC<any> = (props) => {
   const {
     data: {
       contentfulProject: { title },
