+++ src/components/Pages/ProjectsPageComponent.tsx
@@ -4,29 +4,25 @@ import React from 'react';
 // Utils
 import { LocalizedLink } from 'gatsby-theme-i18n';
 import * as siteMetaData from '../../utils/siteMetaData';
-import { WithLayoutProps } from '../../Hoc/hoc.types';
-import { Edge } from './pages-components.types';
 
 // Components
 import withLayout from '../../Hoc/PageWrapper/WithLayout';
 import SEO from '../seo';
 
-const ProjectsPageComponent: React.FC<WithLayoutProps> = (props) => {
+const ProjectsPageComponent: React.FC<any> = (props) => {
   const { data, locale } = props;
-  const informationElements: Array<Edge> = data.allContentfulProject.group.find(
-    (lang: { [key: string]: string }) => lang.fieldValue === locale,
+  const informationElements = data.allContentfulProject.group.find(
+    (lang: any) => lang.fieldValue === locale,
   ).edges;
 
-  const renderInformation = (): Array<React.ReactElement> =>
-    informationElements.map(
-      (edge): React.ReactElement => (
-        <li key={edge.node.id}>
-          <LocalizedLink to={`${siteMetaData.menu[1].slug}${edge.node.slug}`}>
-            {edge.node.shortDescription}
-          </LocalizedLink>
-        </li>
-      ),
-    );
+  const renderInformation = () =>
+    informationElements.map((edge: any) => (
+      <li key={edge.node.id}>
+        <LocalizedLink to={`${siteMetaData.menu[1].slug}${edge.node.slug}`}>
+          {edge.node.shortDescription}
+        </LocalizedLink>
+      </li>
+    ));
 
   return (
     <>
