--- src/components/Pages/pages-components.types.tsx
@@ -1,6 +1,0 @@
-// Utils
-import { ContentfulProject } from '../../utils/site.types';
-
-export type Edge = {
-  node: ContentfulProject;
-};
