+++ src/components/image.tsx
@@ -15,8 +15,7 @@ import Img from 'gatsby-image';
  */
 
 const Image: React.FC = () => {
-  // TODO: Typecheck any
-  const data: { [key: string]: any } = useStaticQuery(graphql`
+  const data = useStaticQuery(graphql`
     query {
       placeholderImage: file(relativePath: { eq: "gatsby-astronaut.png" }) {
         childImageSharp {
