--- src/pages/pages.types.tsx
@@ -1,7 +1,0 @@
-export interface ProjectsPageProps {
-  data: {
-    allContentfulProject: {
-      group: [{ [key: string]: string }];
-    };
-  };
-}
