+++ src/templates/Project/Project.tsx
@@ -2,13 +2,10 @@
 import React from 'react';
 import { graphql } from 'gatsby';
 
-// Utils
-import { ProjectTemplateProps } from '../templates.types';
-
 // Components
 import ProjectPageComponent from '../../components/Pages/ProjectPageComponent';
 
-const Project: React.FC<ProjectTemplateProps> = (props) => {
+const Project: React.FC<any> = (props) => {
   const { data } = props;
   return <ProjectPageComponent data={data} />;
 };
