--- src/templates/templates.types.tsx
@@ -1,5 +1,0 @@
-export interface ProjectTemplateProps {
-  data: {
-    title: string;
-  };
-}
