--- src/utils/site.types.tsx
@@ -1,6 +1,0 @@
-export interface ContentfulProject {
-  title: string;
-  slug: string;
-  shortDescription: string;
-  id: string;
-}
